backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Celery and Redis for task queuing
celery==5.3.4
redis==5.0.1

# LangChain and AI/ML
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1

# LangGraph for agent orchestration
langgraph==0.0.20

# MCP (Model Context Protocol) - these may need to be installed from source
# mcp==0.1.0
# langchain-mcp==0.1.0

# Vector stores and embeddings
chromadb==0.4.18
tiktoken==0.5.2

# GitHub and Git integration
GitPython==3.1.40
PyGithub==2.1.1

# Document processing
python-docx==1.1.0
pypdf2==3.0.1
beautifulsoup4==4.12.2
markdownify==0.11.6

# Data processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Environment and system
python-dotenv==1.0.0
psutil==5.9.6
pathlib==1.0.1

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Validation and serialization
pydantic==2.5.2
SQLAlchemy
psycopg[binary]

# Optional: Docker SDK for Python (for Docker tool integration)
docker==6.1.3

# Optional: Kubernetes client (for k8s tool integration)
kubernetes==28.1.0 